Task#1 Improve local transitions engaging choice pseudostates.
**How can I best achieve it now?**

Check if the code wants to exit the choice pseudostate parent region
again. If yes, omit the region exit code.

**What do I want?**
The generated code should take into account that once the choice
pseudostate is entered, its parent region is in fact exited already. It
doesnt need to be exited again.

**Why do I want it?**
Now the result code is ugly, because when the choice pseudostate
outgoing transition is local type, it executes region exit, which
sometimes is not necessary.