Added separated projects for C sources.
Now there are separate C projects for expected result, generated result
and unit tests of the expected result.

The unit test for the expected result will allow to create the expected
output sources with TDD methodology.

Separate projects for generated code and expected code will allow to
compile them and check if they are running.

Now all projects are in a single repository to maintain consistency
among projects' versions.