Added a simple test class diagram.
The previously used class diagram which was the very first diagram used
in this project, had grown uncontrollably together with the new features
added to the generator, and had become difficult to read. Because of
this, it was accidentally damaged in the meantime when migrating to a
newer version of Papyrus.
This commit introduces a minimal state diagram which illustrates
implementation of the very early features such as polymorphism,
different property types, array members, literal constants and virtual
member functions. The class diagram available in the Papyrus project
represents classes in Tkind package.