Push forward Junction Pseudostates.
Implemented the correct sequence of transition effect behaviors after
all compound transition guards are verified to be true.
Added compound State36 in the Junction test diagram for verification of
the sequence of code execution in case of compound transitions
containing junction pseudostates accross different regions - as
expected, the generated code doesn't call the exit functions of the
source states - must be fixed.